+++ src/main/java/pro/sky/petshelterbot/constants/Commands.java
@@ -1,15 +1,24 @@
 package pro.sky.petshelterbot.constants;
 
 public interface Commands {
+    String ABOUT_SHELTER_INFO = "about_shelter_info";
+    String CALL_VOLUNTEER = "/call_volunteer";
+    String CANCEL_VOLUNTEER_CALL = "/cancel";
+    String CLOSE_DIALOG="/close";
+    String CLOSE_DIALOG_RU="/завершить";
+    String JOIN_DIALOG="/join";
+    String MENU = "/menu";
+    String MENU_RU = "/меню";
+    String OPENING_HOURS_AND_ADDRESS_INFO = "opening_hours_and_address_info";
+    String SECURITY_INFO = "security_info";
+    String SHELTER_CHOICE = "/shelter-";
     String START="/start";
     String RESET_SHELTER="/reset-shelter";
     String RESET_SHELTER_RU="/приют";
-    String SHELTER_CHOICE = "/shelter-";
-    String MENU = "/menu";
-    String MENU_RU = "/меню";
-    String ABOUT_SHELTER_INFO = "about_shelter_info";
-    String OPENING_HOURS_AND_ADDRESS_INFO = "opening_hours_and_address_info";
-    String SECURITY_INFO = "security_info";
 
+    //For next stage development
+    String HAVE_A_BREAK = "/have_a_break";
+    //For next stage development
+    String RESUME_SERVICE = "/resume_service";
 
 }
