--- src/main/java/pro/sky/petshelterbot/constants/DialogCommandNames.java
@@ -1,12 +1,0 @@
-package pro.sky.petshelterbot.constants;
-
-public interface DialogCommandNames extends Commands {
-    String CALL_VOLUNTEER="/call_volunteer";
-    String JOIN_DIALOG="/join_dialog";
-    String CLOSE_DIALOG="/close";
-    String CLOSE_DIALOG_RU="/завершить";
-    //For next stage development
-    String HAVE_A_BREAK = "/have_a_break";
-    //For next stage development
-    String RESUME_SERVICE = "/resume_service";
-}
