+++ src/main/java/pro/sky/petshelterbot/constants/MessageKeys.java
@@ -1,0 +1,5 @@
+package pro.sky.petshelterbot.constants;
+
+public interface MessageKeys {
+   
+}
